# Refactor renderer to allow overloading of concatenation