--- a/lib/marked.js
+++ b/lib/marked.js
@@ -560,7 +560,7 @@ InlineLexer.output = function(src, links, options) {
  */
 
 InlineLexer.prototype.output = function(src) {
-  var out = ''
+  var out = []
     , link
     , text
     , href
@@ -570,7 +570,7 @@ InlineLexer.prototype.output = function(src) {
     // escape
     if (cap = this.rules.escape.exec(src)) {
       src = src.substring(cap[0].length);
-      out += cap[1];
+      out.push(cap[1]);
       continue;
     }
 
@@ -586,7 +586,7 @@ InlineLexer.prototype.output = function(src) {
         text = escape(cap[1]);
         href = text;
       }
-      out += this.renderer.link(href, null, text);
+      out.push(this.renderer.link(href, null, text));
       continue;
     }
 
@@ -595,7 +595,7 @@ InlineLexer.prototype.output = function(src) {
       src = src.substring(cap[0].length);
       text = escape(cap[1]);
       href = text;
-      out += this.renderer.link(href, null, text);
+      out.push(this.renderer.link(href, null, text));
       continue;
     }
 
@@ -607,11 +607,11 @@ InlineLexer.prototype.output = function(src) {
         this.inLink = false;
       }
       src = src.substring(cap[0].length);
-      out += this.options.sanitize
+      out.push(this.options.sanitize
         ? this.options.sanitizer
           ? this.options.sanitizer(cap[0])
           : escape(cap[0])
-        : cap[0]
+        : cap[0])
       continue;
     }
 
@@ -619,10 +619,10 @@ InlineLexer.prototype.output = function(src) {
     if (cap = this.rules.link.exec(src)) {
       src = src.substring(cap[0].length);
       this.inLink = true;
-      out += this.outputLink(cap, {
+      out.push(this.outputLink(cap, {
         href: cap[2],
         title: cap[3]
-      });
+      }));
       this.inLink = false;
       continue;
     }
@@ -634,12 +634,12 @@ InlineLexer.prototype.output = function(src) {
       link = (cap[2] || cap[1]).replace(/\s+/g, ' ');
       link = this.links[link.toLowerCase()];
       if (!link || !link.href) {
-        out += cap[0].charAt(0);
+        out.push(cap[0].charAt(0));
         src = cap[0].substring(1) + src;
         continue;
       }
       this.inLink = true;
-      out += this.outputLink(cap, link);
+      out.push(this.outputLink(cap, link));
       this.inLink = false;
       continue;
     }
@@ -647,42 +647,42 @@ InlineLexer.prototype.output = function(src) {
     // strong
     if (cap = this.rules.strong.exec(src)) {
       src = src.substring(cap[0].length);
-      out += this.renderer.strong(this.output(cap[2] || cap[1]));
+      out.push(this.renderer.strong(this.output(cap[2] || cap[1])));
       continue;
     }
 
     // em
     if (cap = this.rules.em.exec(src)) {
       src = src.substring(cap[0].length);
-      out += this.renderer.em(this.output(cap[2] || cap[1]));
+      out.push(this.renderer.em(this.output(cap[2] || cap[1])));
       continue;
     }
 
     // code
     if (cap = this.rules.code.exec(src)) {
       src = src.substring(cap[0].length);
-      out += this.renderer.codespan(escape(cap[2], true));
+      out.push(this.renderer.codespan(escape(cap[2], true)));
       continue;
     }
 
     // br
     if (cap = this.rules.br.exec(src)) {
       src = src.substring(cap[0].length);
-      out += this.renderer.br();
+      out.push(this.renderer.br());
       continue;
     }
 
     // del (gfm)
     if (cap = this.rules.del.exec(src)) {
       src = src.substring(cap[0].length);
-      out += this.renderer.del(this.output(cap[1]));
+      out.push(this.renderer.del(this.output(cap[1])));
       continue;
     }
 
     // text
     if (cap = this.rules.text.exec(src)) {
       src = src.substring(cap[0].length);
-      out += this.renderer.text(escape(this.smartypants(cap[0])));
+      out.push(this.renderer.text(escape(this.smartypants(cap[0]))));
       continue;
     }
 
@@ -692,7 +692,7 @@ InlineLexer.prototype.output = function(src) {
     }
   }
 
-  return out;
+  return this.renderer.combine(out);
 };
 
 /**
@@ -760,6 +760,10 @@ InlineLexer.prototype.mangle = function(text) {
 function Renderer(options) {
   this.options = options || {};
 }
+
+Renderer.prototype.combine = function(out) {
+  return out.join('');
+};
 
 Renderer.prototype.code = function(code, lang, escaped) {
   if (this.options.highlight) {
@@ -930,12 +934,12 @@ Parser.prototype.parse = function(src) {
   this.inline = new InlineLexer(src.links, this.options, this.renderer);
   this.tokens = src.reverse();
 
-  var out = '';
+  var out = [];
   while (this.next()) {
-    out += this.tok();
+    out.push(this.tok());
   }
 
-  return out;
+  return this.renderer.combine(out);
 };
 
 /**
@@ -973,6 +977,8 @@ Parser.prototype.parseText = function() {
  */
 
 Parser.prototype.tok = function() {
+  var body = [];
+
   switch (this.token.type) {
     case 'space': {
       return '';
@@ -993,15 +999,12 @@ Parser.prototype.tok = function() {
     }
     case 'table': {
       var header = ''
-        , body = ''
-        , i
-        , row
-        , cell
-        , flags
-        , j;
+      , i
+      , row
+      , cell = ''
+      , flags
+      , j;
 
-      // header
-      cell = '';
       for (i = 0; i < this.token.header.length; i++) {
         flags = { header: true, align: this.token.align[i] };
         cell += this.renderer.tablecell(
@@ -1022,48 +1025,41 @@ Parser.prototype.tok = function() {
           );
         }
 
-        body += this.renderer.tablerow(cell);
+        body.push(this.renderer.tablerow(cell));
       }
-      return this.renderer.table(header, body);
+      return this.renderer.table(header, this.renderer.combine(body));
     }
     case 'blockquote_start': {
-      var body = '';
-
       while (this.next().type !== 'blockquote_end') {
-        body += this.tok();
+        body.push(this.tok());
       }
 
-      return this.renderer.blockquote(body);
+      return this.renderer.blockquote(this.renderer.combine(body));
     }
     case 'list_start': {
-      var body = ''
-        , ordered = this.token.ordered;
+      var ordered = this.token.ordered;
 
       while (this.next().type !== 'list_end') {
-        body += this.tok();
+        body.push(this.tok());
       }
 
-      return this.renderer.list(body, ordered);
+      return this.renderer.list(this.renderer.combine(body), ordered);
     }
     case 'list_item_start': {
-      var body = '';
-
       while (this.next().type !== 'list_item_end') {
-        body += this.token.type === 'text'
+        body.push(this.token.type === 'text'
           ? this.parseText()
-          : this.tok();
+          : this.tok());
       }
-
-      return this.renderer.listitem(body);
     }
-    case 'loose_item_start': {
-      var body = '';
 
+      return this.renderer.listitem(this.renderer.combine(body));
+    case 'loose_item_start': {
       while (this.next().type !== 'list_item_end') {
-        body += this.tok();
+        body.push(this.tok());
       }
 
-      return this.renderer.listitem(body);
+      return this.renderer.listitem(this.renderer.combine(body));
     }
     case 'html': {
       var html = !this.token.pre && !this.options.pedantic