--- a/lib/marked.js
+++ b/lib/marked.js
@@ -583,7 +583,7 @@ InlineLexer.prototype.output = function(src) {
           : this.mangle(cap[1]);
         href = this.mangle('mailto:') + text;
       } else {
-        text = escape(cap[1]);
+        text = this.renderer.escape(cap[1]);
         href = text;
       }
       out.push(this.renderer.link(href, null, text));
@@ -593,7 +593,7 @@ InlineLexer.prototype.output = function(src) {
     // url (gfm)
     if (!this.inLink && (cap = this.rules.url.exec(src))) {
       src = src.substring(cap[0].length);
-      text = escape(cap[1]);
+      text = this.renderer.escape(cap[1]);
       href = text;
       out.push(this.renderer.link(href, null, text));
       continue;
@@ -610,7 +610,7 @@ InlineLexer.prototype.output = function(src) {
       out.push(this.options.sanitize
         ? this.options.sanitizer
           ? this.options.sanitizer(cap[0])
-          : escape(cap[0])
+          : this.renderer.escape(cap[0])
         : cap[0])
       continue;
     }
@@ -661,7 +661,7 @@ InlineLexer.prototype.output = function(src) {
     // code
     if (cap = this.rules.code.exec(src)) {
       src = src.substring(cap[0].length);
-      out.push(this.renderer.codespan(escape(cap[2], true)));
+      out.push(this.renderer.codespan(this.renderer.escape(cap[2], true)));
       continue;
     }
 
@@ -682,7 +682,7 @@ InlineLexer.prototype.output = function(src) {
     // text
     if (cap = this.rules.text.exec(src)) {
       src = src.substring(cap[0].length);
-      out.push(this.renderer.text(escape(this.smartypants(cap[0]))));
+      out.push(this.renderer.text(this.renderer.escape(this.smartypants(cap[0]))));
       continue;
     }
 
@@ -700,12 +700,12 @@ InlineLexer.prototype.output = function(src) {
  */
 
 InlineLexer.prototype.outputLink = function(cap, link) {
-  var href = escape(link.href)
-    , title = link.title ? escape(link.title) : null;
+  var href = this.renderer.escape(link.href)
+    , title = link.title ? this.renderer.escape(link.title) : null;
 
   return cap[0].charAt(0) !== '!'
     ? this.renderer.link(href, title, this.output(cap[1]))
-    : this.renderer.image(href, title, escape(cap[1]));
+    : this.renderer.image(href, title, this.renderer.escape(cap[1]));
 };
 
 /**
@@ -765,6 +765,15 @@ Renderer.prototype.combine = function(out) {
   return out.join('');
 };
 
+Renderer.prototype.escape = function(html, encode) {
+  return html
+    .replace(!encode ? /&(?!#?\w+;)/g : /&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;')
+    .replace(/'/g, '&#39;');
+}
+
 Renderer.prototype.code = function(code, lang, escaped) {
   if (this.options.highlight) {
     var out = this.options.highlight(code, lang);
@@ -776,15 +785,15 @@ Renderer.prototype.code = function(code, lang, escaped) {
 
   if (!lang) {
     return '<pre><code>'
-      + (escaped ? code : escape(code, true))
+      + (escaped ? code : this.escape(code, true))
       + '\n</code></pre>';
   }
 
   return '<pre><code class="'
     + this.options.langPrefix
-    + escape(lang, true)
+    + this.escape(lang, true)
     + '">'
-    + (escaped ? code : escape(code, true))
+    + (escaped ? code : this.escape(code, true))
     + '\n</code></pre>\n';
 };
 
@@ -1080,15 +1089,6 @@ Parser.prototype.tok = function() {
  * Helpers
  */
 
-function escape(html, encode) {
-  return html
-    .replace(!encode ? /&(?!#?\w+;)/g : /&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#39;');
-}
-
 function unescape(html) {
   return html.replace(/&([#\w]+);/g, function(_, n) {
     n = n.toLowerCase();
@@ -1216,7 +1216,7 @@ function marked(src, opt, callback) {
     e.message += '\nPlease report this to https://github.com/chjj/marked.';
     if ((opt || marked.defaults).silent) {
       return '<p>An error occured:</p><pre>'
-        + escape(e.message + '', true)
+        + this.renderer.escape(e.message + '', true)
         + '</pre>';
     }
     throw e;