- Currently, the renderer assumes String concatenation, but to support other methods of rendering (like React), implement an overloadable function (`combine`) in the renderer for concatenation.
- Passes existing tests, except test 10 which is already failing on master.

I'm opening this PR merely mostly for discussion since it negatively affects performance, and in case anyone else is looking to use React with marked. 

Performance characteristics of this change:

```
# before
marked completed in 2799ms.
marked (gfm) completed in 2823ms.
marked (pedantic) completed in 2449ms.

# after
marked completed in 3291ms.
marked (gfm) completed in 3400ms.
marked (pedantic) completed in 2937ms.
```

Also, [here's an example of a React](https://gist.github.com/mbranch/1f9493b9346bdbfce6fe) renderer that works with this change.