@mina86 2015-10-07
20% slow down sounds like a steep price to pay.