@joshbruce 2018-01-21
Closing as stale, merge conflicts, missing tests (??), and performance hit too steep for functionality.